split resource admin tests
