+++ .github/workflows/designer-dotnet-test.yaml
@@ -36,4 +36,4 @@ jobs:
            dotnet build backend/Designer.sln -v m
       - name: Test
         run: |
-           dotnet test backend/Designer.sln --filter Category!=GiteaIntegrationTest -v n --no-build --blame-hang-timeout 1min
+           dotnet test backend/Designer.sln --filter Category!=GiteaIntegrationTest -v n --no-build --blame-hang-timeout 3min
