+++ .github/workflows/dotnet-tests-attempt.yaml
@@ -21,4 +21,4 @@ jobs:
       - uses: actions/checkout@v3
       - name: Test
         run: |
-           dotnet test backend/Designer.sln --filter Category!=GiteaIntegrationTest -v n --blame-hang-timeout 1min
+           dotnet test backend/Designer.sln --filter Category!=GiteaIntegrationTest -v n --blame-hang-timeout 3min
