+++ backend/tests/Designer.Tests/Controllers/ResourceAdminController/GetPublishStatusByIdTests.cs
@@ -1,0 +1,58 @@
+﻿using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Altinn.Studio.Designer.Enums;
+using Altinn.Studio.Designer.Models;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Moq;
+using Xunit;
+
+namespace Designer.Tests.Controllers.ResourceAdminController
+{
+    public class GetPublishStatusByIdTests : ResourceAdminControllerTestsBaseClass<GetPublishStatusByIdTests>
+    {
+        public GetPublishStatusByIdTests(WebApplicationFactory<Altinn.Studio.Designer.Controllers.ResourceAdminController> factory) : base(factory)
+        {
+        }
+
+        [Fact]
+        public async Task GetResourceStatusById_Passing_Repository_OK()
+        {
+            // Arrange
+            string uri = $"{VersionPrefix}/ttd/resources/publishstatus/ttd-resources/ttd_testresource";
+
+            RepositoryMock
+                .Setup(r => r.GetServiceResourceById(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()))
+                .Returns(
+                    new ServiceResource
+                    {
+                        Identifier = "testresource",
+                        Title = new Dictionary<string, string>(),
+                        Description = new Dictionary<string, string>(),
+                        RightDescription = new Dictionary<string, string>(),
+                        Homepage = "test.no",
+                        Status = string.Empty,
+                        IsPartOf = string.Empty,
+                        ThematicArea = string.Empty,
+                        ResourceReferences = GetTestResourceReferences(),
+                        Delegable = true,
+                        Visible = true,
+                        Version = "2023.12",
+                        HasCompetentAuthority = new CompetentAuthority { Organization = "ttd", Orgcode = "test", Name = new Dictionary<string, string>() },
+                        Keywords = GetTestKeywords(),
+                        ResourceType = ResourceType.Default,
+                    });
+
+            using HttpRequestMessage httpRequestMessage = new HttpRequestMessage(HttpMethod.Get, uri);
+
+            // Act
+            using HttpResponseMessage res = await HttpClient.Value.SendAsync(httpRequestMessage).ConfigureAwait(false);
+
+            // Assert
+            Assert.Equal(HttpStatusCode.OK, res.StatusCode);
+
+        }
+
+    }
+}
